requirements.txt
# Web framework
Flask>=2.3,<3.0
Flask-SQLAlchemy>=3.0,<4.0
Flask-Migrate>=4.0,<5.0

# Async / background tasks
celery>=5.5,<6.0
redis>=5.0,<6.0

# Database driver
psycopg2-binary>=2.9,<3.0

# CSV processing
pandas>=2.1,<3.0
chardet>=5.2,<6.0

# SSE / CORS / utils
Flask-Cors>=3.1

# Environment variables
python-dotenv>=1.0,<2.0

# Optional helpers
requests>=2.31,<3.0

# Production server
gunicorn>=21.2,<22.0

# Testing
pytest>=7.4,<8.0
pytest-mock>=3.12,<4.0
httpx>=0.26,<1.0
pytest-flask>=1.3,<2.0
